requirements.txt
absl-py==2.0.0
argcomplete==3.2.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cppcuda-tutorial==1.0
cycler==0.12.1
e==1.4.5
et-xmlfile==1.1.0
evo==1.26.0
fonttools==4.47.0
fused-conv-random-k==0.0.0
fused-conv-select-k==0.0.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
KNN-CUDA==0.2
lz4==4.3.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.4
natsort==8.4.0
ninja==1.11.1
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pip==23.2.1
pointpillars==0.0.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rosbags==0.9.19
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.2
setuptools==68.0.0
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==1.13.1+cu117
torch-summary==1.4.5
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.6
Werkzeug==3.0.0
wheel==0.41.2
zipp==3.17.0
zstandard==0.22.0
